\name{GeneCycle-internal}
\alias{dominant.freqs.single}
\alias{periodogram.spec.single}
\alias{periodogram.spec}
\alias{periodogram.freq}
\alias{fisher.g.test.single}
\alias{is.constant.single}

\alias{robust.spectrum.single}
\alias{spearman}
\alias{gPopCreate}
\alias{myresample}

\alias{fit.DC}
\alias{fit.freqs}
\alias{fit.pi}
\alias{g.statistic}

\title{Internal GeneCycle Functions}
\description{
  Internal GeneCycle functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
